Implement new records resource.
 Currently the resource path is `/_records` because the old records resource is still in use by e-petitions.

 Once e-petitions starts using the new resource we will remove the old resource and change the path to `/records`.

  It was required to start working on register resource. register resource uses registers.yaml, fields.yaml file whose structure should match with the response of new record resource. Now we can grep the response and just copy in these files in every environment with proper data.
